Extract LTL mutator into standalone pip-installable sub-package
Factor out the core mutation logic (codebook, syntacticmutator, ltlnode,
spotutils, ANTLR grammar) into a new `ltlmutator/` package at the repo
root so it can be reused independently (closes #142).

The ltltoeng circular dependency is broken via lazy import — the package
works standalone without ltltoeng/inflect/wordfreq, falling back to
str() for __to_english__(). All consumers updated to import from
ltlmutator.
