+++ experiments/benchmark_evaluator.py
@@ -17,8 +17,8 @@ import time
 # Add src to path for LTL parsing and equivalence checking
 sys.path.insert(0, str(Path(__file__).parent.parent / 'src'))
 
-from ltlnode import parse_ltl_string
-from spotutils import areEquivalent
+from ltlmutator import parse_ltl_string
+from ltlmutator.spotutils import areEquivalent
 import spot
