+++ ltlmutator/src/ltlmutator.egg-info/PKG-INFO
@@ -1,0 +1,6 @@
+Metadata-Version: 2.4
+Name: ltlmutator
+Version: 0.1.0
+Summary: LTL formula semantic and syntactic mutation library
+Requires-Python: >=3.9
+Requires-Dist: antlr4-python3-runtime>=4.13.1
