+++ ltlmutator/src/ltlmutator.egg-info/requires.txt
@@ -1,0 +1,1 @@
+antlr4-python3-runtime>=4.13.1
