+++ requirements.txt
@@ -33,3 +33,4 @@ psycopg2-binary
 flask-login
 inflect
 wordfreq>=3.0
+./ltlmutator
